Fixed error with iterators looping forever
The problem was that the copy of the iterator has been used for
iteration. This made it so that if there were more than 1 component in
the archetype it would loop forever. We did not see it yet and i am not
sure why as we already have systems that loop over every button for
example. The problematic code was both in zig and lua code but the
problem only happened when we added commands for some reason and only on
zig side.